ai_summarizer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-legal-document-summarizer"
version = "0.1.0"
description = "Summarization, clause extraction, and risk checks for legal documents."
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
  "striprtf>=0.0.26",
  "transformers>=4.40.0",
  "torch>=2.1.0"
]

[project.scripts]
legal-ai = "ai_summarizer.cli:main"
